% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpiechartclustermarkers.R
\name{addpiechartclustermarkers2}
\alias{addpiechartclustermarkers2}
\title{Change leaflet cluster markers to pie charts}
\usage{
addpiechartclustermarkers2(map, .data, .colors, group, clusterId = "toto", ...)
}
\arguments{
\item{map}{the map to add awesome pie chart cluster markers to}

\item{.data}{data for the cluster markers}

\item{.colors}{a vector of colors of at least the same size that nlevels(.data[[group]])}

\item{group}{the name of a factor variable of .data}

\item{...}{further arguments to be passed to addAwesomeMarkers}
}
\description{
Change leaflet cluster markers to pie charts
}
\examples{
library(leaflet)
library(dplyr)
data("breweries91",package="leaflet")
breweries91$goodbeer<-sample(as.factor(c("terrific","marvelous","culparterretaping")),nrow(breweries91),replace=T)
leaflet(breweries91) \%>\%addTiles()\%>\%addpiechartclustermarkers2(.data=breweries91,.colors=c("red","green","blue"),group="goodbeer")
}
